requirements.txt
# requirements.txt
apache-airflow==2.7.1
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.1.0
pytest==7.4.0
faker==19.3.0
